Refactor FeatureToggles packages in debug menu
