+++ modules/app/src/main/java/kekmech/ru/mpeiapp/di/AppModule.kt
@@ -106,8 +106,5 @@ private class RemoteConfigWrapperImpl : RemoteConfigWrapper {
     private val remoteConfig
         get() = FirebaseRemoteConfig.getInstance()
 
-    @Deprecated("Use `getUntyped` instead", replaceWith = ReplaceWith("getUntyped"))
-    override fun get(featureToggleKey: String): Boolean = remoteConfig.getBoolean(featureToggleKey)
-
     override fun getUntyped(name: String): String = remoteConfig.getString(name)
 }
