+++ modules/common/feature_toggles/src/main/kotlin/kekmech/ru/common_feature_toggles/RemoteConfigWrapper.kt
@@ -2,8 +2,5 @@ package kekmech.ru.common_feature_toggles
 
 interface RemoteConfigWrapper {
 
-    @Deprecated("Use `getUntyped` instead", replaceWith = ReplaceWith("getUntyped"))
-    operator fun get(featureToggleKey: String): Boolean
-
     fun getUntyped(name: String): String
 }
