+++ modules/common/feature_toggles/src/main/kotlin/kekmech/ru/common_feature_toggles/RemoteVariable.kt
@@ -13,7 +13,7 @@ sealed class RemoteVariable<T : Any>(
 ) {
 
     val value get() =
-        valueHolder[name]
+        valueHolder.get(name)
             ?.let(deserializationStrategy)
             ?: defaultValue
 }
