+++ modules/common/feature_toggles/src/main/kotlin/kekmech/ru/common_feature_toggles/RemoteVariableValueHolder.kt
@@ -1,9 +1,20 @@
 package kekmech.ru.common_feature_toggles
 
-interface RemoteVariableValueHolder {
+class RemoteVariableValueHolder internal constructor(
+    private val remoteConfigWrapper: RemoteConfigWrapper,
+    private val middleware: Middleware?,
+) {
 
     /**
      * Get serialized variable value
      */
-    operator fun get(name: String): String?
+    fun get(name: String): String? =
+        middleware?.get(name)
+            ?: remoteConfigWrapper.getUntyped(name)
+                .takeIf { it.isNotEmpty() }
+
+    interface Middleware {
+
+        fun get(name: String): String?
+    }
 }
